rename: 폴더 구조 변경에 따른 import 경로 수정 (CC-174)
